+++ src/Components/Navbar/Navbar-comp.jsx
@@ -29,4 +29,4 @@ const Navbar = () => {
   )
 }
 
-export default Navbar;
+export {Navbar};
