+++ src/Components/index.js
@@ -1,0 +1,7 @@
+export {Navbar} from './Navbar/Navbar-comp';
+export {Product} from "./Product/Product";
+export {Login} from './Authentication/Login';
+export { Cart } from "./Cart/Cart"
+export {Wishlist} from "./Wishlist/Wishlist";
+export {Home} from "./Screen/Home"
+export {Filter} from "./Filter/Filter"
